+++ components/stsdk/caps_battery.h
@@ -18,10 +18,6 @@
 
 #include "caps/iot_caps_helper_battery.h"
 
-#ifdef __cplusplus
-extern "C" {
-#endif
-
 typedef struct caps_battery_data {
     IOT_CAP_HANDLE* handle;
     void *usr_data;
@@ -40,7 +36,4 @@ typedef struct caps_battery_data {
 } caps_battery_data_t;
 
 caps_battery_data_t *caps_battery_initialize(IOT_CTX *ctx, const char *component, void *init_usr_cb, void *usr_data);
-#ifdef __cplusplus
-}
-#endif