+++ main/ad2_uart_cli.h
@@ -27,10 +27,6 @@
 #define PROMPT_STRING "AD2IOT # "
 #define AD2_HELP_CMD "help"
 
-#ifdef __cplusplus
-extern "C" {
-#endif
-
 typedef void (* command_function_t)(const char *string);
 
 typedef struct cli_command {
@@ -47,7 +43,4 @@ typedef struct cli_command_list {
 void uart_cli_main();
 void cli_register_command(cli_cmd_t* cmd);
 
-#ifdef __cplusplus
-}
-#endif
 #endif /* _AD2_UART_CLI_H_ */